Added support for genetic algorithm in DiCE
