+++ dice_ml/data_interfaces/public_data_interface.py
@@ -271,6 +271,7 @@ class PublicData:
                 elif colidx not in encoded_cats_ixs and col in features_to_vary:
                     ixs.append(colidx)
             return ixs
+
     def from_label(self, data):
         out = data.copy()
         for column in self.categorical_feature_names:
@@ -321,6 +322,7 @@ class PublicData:
                 data = pd.DataFrame(data=data, index=index,
                                     columns=self.encoded_feature_names)
                 return self.from_dummies(data)
+
             elif encoding == 'label':
                 data = pd.DataFrame(data=data, index=index,
                                     columns=self.feature_names)
@@ -368,17 +370,18 @@ class PublicData:
 
         test = test.reset_index(drop=True)
 
-        if encoding == 'one-hot':
+        if encoding == 'label':
+            for column in self.categorical_feature_names:
+                test[column] = self.labelencoder[column].transform(test[column])
+            return self.normalize_data(test)
+
+        elif encoding == 'one-hot':
             temp = self.prepare_df_for_encoding()
             temp = temp.append(test, ignore_index=True, sort=False)
             temp = self.one_hot_encode_data(temp)
             temp = self.normalize_data(temp)
-            return temp.tail(test.shape[0]).reset_index(drop=True)
 
-        elif encoding == 'label':
-            for column in self.categorical_feature_names:
-                test[column] = self.labelencoder[column].transform(test[column])
-            return self.normalize_data(test)
+            return temp.tail(test.shape[0]).reset_index(drop=True)
 
     def get_dev_data(self, model_interface, desired_class, filter_threshold=0.5):
         """Constructs dev data by extracting part of the test data for which finding counterfactuals make sense."""
