+++ dice_ml/explainer_interfaces/dice_genetic.py
@@ -44,11 +44,7 @@ class DiceGenetic(ExplainerBase):
     def generate_counterfactuals(self, query_instance, total_CFs, desired_class="opposite", proximity_weight=0.5,
                                  diversity_weight=1.0, categorical_penalty=0.1, algorithm="DiverseCF",
                                  features_to_vary="all", permitted_range=None, yloss_type="l2_loss",
-                                 diversity_loss_type="dpp_style:inverse_dist", feature_weights="inverse_mad",
-                                 optimizer="tensorflow:adam", learning_rate=0.05, min_iter=500, max_iter=5000,
-                                 project_iter=0, loss_diff_thres=1e-5, loss_converge_maxiter=1, verbose=False,
-                                 init_near_query_instance=True, tie_random=False, stopping_threshold=0.5,
-                                 posthoc_sparsity_param=0.1, posthoc_sparsity_algorithm="linear"):
+                                 diversity_loss_type="dpp_style:inverse_dist", feature_weights="inverse_mad", stopping_threshold=0.5, posthoc_sparsity_param=0.1, posthoc_sparsity_algorithm="linear"):
         """Generates diverse counterfactual explanations
 
         :param query_instance: A dictionary of feature names and values. Test point of interest.
@@ -61,20 +57,11 @@ class DiceGenetic(ExplainerBase):
 
         :param algorithm: Counterfactual generation algorithm. Either "DiverseCF" or "RandomInitCF".
         :param features_to_vary: Either a string "all" or a list of feature names to vary.
-        param permitted_range: Dictionary with continuous feature names as keys and permitted min-max range in list as values. Defaults to the range inferred from training data. If None, uses the parameters initialized in data_interface.
+        :param permitted_range: Dictionary with continuous feature names as keys and permitted min-max range in list as values. Defaults to the range inferred from training data. If None, uses the parameters initialized in data_interface.
         :param yloss_type: Metric for y-loss of the optimization function. Takes "l2_loss" or "log_loss" or "hinge_loss".
         :param diversity_loss_type: Metric for diversity loss of the optimization function. Takes "avg_dist" or "dpp_style:inverse_dist".
         :param feature_weights: Either "inverse_mad" or a dictionary with feature names as keys and corresponding weights as values. Default option is "inverse_mad" where the weight for a continuous feature is the inverse of the Median Absolute Devidation (MAD) of the feature's values in the training set; the weight for a categorical feature is equal to 1 by default.
-        :param optimizer: Tensorflow optimization algorithm. Currently tested only with "tensorflow:adam".
-        :param learning_rate: Learning rate for optimizer.
-        :param min_iter: Min iterations to run gradient descent for.
-        :param max_iter: Max iterations to run gradient descent for.
-        :param project_iter: Project the gradients at an interval of these many iterations.
-        :param loss_diff_thres: Minimum difference between successive loss values to check convergence.
-        :param loss_converge_maxiter: Maximum number of iterations for loss_diff_thres to hold to declare convergence. Defaults to 1, but we assigned a more conservative value of 2 in the paper.
-        :param verbose: Print intermediate loss value.
-        :param init_near_query_instance: Boolean to indicate if counterfactuals are to be initialized near query_instance.
-        :param tie_random: Used in rounding off CFs and intermediate projection.
+
         :param stopping_threshold: Minimum threshold for counterfactuals target class probability.
         :param posthoc_sparsity_param: Parameter for the post-hoc operation on continuous features to enhance sparsity.
         :param posthoc_sparsity_algorithm: Perform either linear or binary search. Takes "linear" or "binary". Prefer binary search when a feature range is large (for instance, income varying from 10k to 1000k) and only if the features share a monotonic relationship with predicted outcome in the model.
@@ -86,92 +73,14 @@ class DiceGenetic(ExplainerBase):
         self.check_permitted_range(permitted_range)
         self.do_param_initializations(total_CFs, algorithm, features_to_vary, yloss_type, diversity_loss_type, feature_weights, proximity_weight, diversity_weight, categorical_penalty)
 
-        query_instance, test_pred = self.find_counterfactuals(query_instance, desired_class, optimizer, learning_rate,
-                                                              min_iter, max_iter, project_iter, loss_diff_thres,
-                                                              loss_converge_maxiter, verbose, init_near_query_instance,
-                                                              tie_random, stopping_threshold, posthoc_sparsity_param,
-                                                              posthoc_sparsity_algorithm)
-        return exp.CounterfactualExamples(self.data_interface, query_instance,
-        test_pred, self.final_cfs, self.cfs_preds, self.final_cfs_sparse, self.cfs_preds_sparse, posthoc_sparsity_param, desired_class, encoding='label')
+        query_instance, test_pred = self.find_counterfactuals(query_instance, desired_class, stopping_threshold, posthoc_sparsity_param, posthoc_sparsity_algorithm)
+        return exp.CounterfactualExamples(self.data_interface, query_instance, test_pred, self.final_cfs, self.cfs_preds, self.final_cfs_sparse, self.cfs_preds_sparse, posthoc_sparsity_param, desired_class, encoding='label')
 
     def predict_fn(self, input_instance):
         """prediction function"""
         temp_preds = self.model.get_output(input_instance).numpy()
         return np.array([preds[(self.num_output_nodes - 1):] for preds in temp_preds], dtype=np.float32)
 
-    def do_cf_initializations(self, total_CFs, algorithm, features_to_vary):
-        """Intializes CFs and other related variables."""
-
-        self.cf_init_weights = [total_CFs, algorithm, features_to_vary]
-
-        if algorithm == "RandomInitCF":
-            # no. of times to run the experiment with random inits for diversity
-            self.total_random_inits = total_CFs
-            self.total_CFs = 1  # size of counterfactual set
-        else:
-            self.total_random_inits = 0
-            self.total_CFs = total_CFs  # size of counterfactual set
-
-        # freeze those columns that need to be fixed
-        if features_to_vary != self.features_to_vary:
-            self.features_to_vary = features_to_vary
-            self.feat_to_vary_idxs = self.data_interface.get_indexes_of_features_to_vary(
-                features_to_vary=features_to_vary)
-            self.freezer = [1.0 if ix in self.feat_to_vary_idxs else 0.0 for ix in range(len(self.minx[0]))]
-
-        # CF initialization
-        if len(self.cfs) != self.total_CFs:
-            self.cfs = []
-            for kx in range(self.population_size):
-                self.temp_cfs = []
-                for ix in range(self.total_CFs):
-                    one_init = [[]]
-                    for jx in range(len(self.data_interface.feature_names)):
-                        one_init[0].append(np.random.uniform(self.minx[0][jx], self.maxx[0][jx]))
-                    self.temp_cfs.append(np.array(one_init))
-                self.cfs.append(self.temp_cfs)
-
-    def do_loss_initializations(self, yloss_type, diversity_loss_type, feature_weights, encoded=True):
-        """Intializes variables related to main loss function"""
-
-        self.loss_weights = [yloss_type, diversity_loss_type, feature_weights]
-        # define the loss parts
-        self.yloss_type = yloss_type
-        self.diversity_loss_type = diversity_loss_type
-
-        # define feature weights
-        if feature_weights != self.feature_weights_input:
-            self.feature_weights_input = feature_weights
-            if feature_weights == "inverse_mad":
-                normalized_mads = self.data_interface.get_valid_mads(normalized=True)
-                feature_weights = {}
-                for feature in normalized_mads:
-                    feature_weights[feature] = round(1 / normalized_mads[feature], 2)
-
-            feature_weights_list = []
-            if(encoded):
-                for feature in self.data_interface.encoded_feature_names:
-                    if feature in feature_weights:
-                        feature_weights_list.append(feature_weights[feature])
-                    else:
-                        feature_weights_list.append(1.0)
-            else:
-                for feature in self.data_interface.feature_names:
-                    if feature in feature_weights:
-                        feature_weights_list.append(feature_weights[feature])
-                    else:
-                        feature_weights_list.append(self.data_interface.label_encoded_data[feature].max())
-            self.feature_weights_list = [feature_weights_list]
-
-    def update_hyperparameters(self, proximity_weight, diversity_weight, categorical_penalty):
-        """Update hyperparameters of the loss function"""
-
-        self.hyperparameters = [proximity_weight, diversity_weight, categorical_penalty]
-        self.proximity_weight = proximity_weight
-        self.diversity_weight = diversity_weight
-        self.categorical_penalty = categorical_penalty
-
-    # @property
     def compute_yloss(self, cfs):
         """Computes the first part (y-loss) of the loss function."""
         yloss = 0.0
@@ -279,26 +188,21 @@ class DiceGenetic(ExplainerBase):
                 # random probability
                 prob = random.random()
 
-                # if prob is less than 0.45, insert gene
-                # from parent 1
+                # if prob is less than 0.45, insert gene from parent 1
                 if prob < 0.45:
                     one_init[0].append(gp1)
 
-                    # if prob is between 0.45 and 0.90, insert
-                # gene from parent 2
+                # if prob is between 0.45 and 0.90, insert gene from parent 2
                 elif prob < 0.90:
                     one_init[0].append(gp2)
 
-                    # otherwise insert random gene(mutate),
-                # for maintaining diversity
+                #otherwise insert random gene(mutate) for maintaining diversity
                 else:
                      one_init[0].append(np.random.uniform(self.minx[0][jx], self.maxx[0][jx]))
             child_chromosome.append(np.array(one_init))
         return child_chromosome
 
-    def find_counterfactuals(self, query_instance, desired_class, optimizer, learning_rate, min_iter, max_iter,
-                             project_iter, loss_diff_thres, loss_converge_maxiter, verbose, init_near_query_instance,
-                             tie_random, stopping_threshold, posthoc_sparsity_param, posthoc_sparsity_algorithm):
+    def find_counterfactuals(self, query_instance, desired_class, stopping_threshold, posthoc_sparsity_param, posthoc_sparsity_algorithm):
         """Finds counterfactuals by gradient-descent."""
 
         # Prepares user defined query_instance for DiCE.
@@ -319,7 +223,6 @@ class DiceGenetic(ExplainerBase):
         elif self.target_cf_class == 1 and self.stopping_threshold < 0.5:
             self.stopping_threshold = 0.75
 
-        current_best_loss = np.inf
         population = self.cfs.copy()
 
         start_time = timeit.default_timer()
@@ -342,11 +245,12 @@ class DiceGenetic(ExplainerBase):
                 self.valid_cfs_found = True
                 break
 
+            # 10% of the next generation is fittest members of current generation
             population_fitness = sorted(population_fitness, key=lambda x: x[1])
             s = int((10 * self.population_size) / 100)
-
             new_generation = [population[tup[0]] for tup in population_fitness[:s]]
 
+            # 90% of the next generation obtained from top 50% of fittest members of current generation
             s = int((90 * self.population_size) / 100)
             for _ in range(s):
                 parent1 = random.choice(population[:int(50 * self.population_size / 100)])
